server/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@47b9ac10710dd51e6e4e1304223d6dc543c47a54
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
bitsandbytes==0.39.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cohere==4.40
coloredlogs==15.0.1
colorlog==6.8.0
contextlib2==21.6.0
crewai==0.1.17
curl-cffi==0.5.10
dataclasses-json==0.6.3
datasets==2.16.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
duckduckgo_search==4.1.1
einops==0.6.1
evaluate==0.4.0
fastapi==0.108.0
fastavro==1.9.2
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.4
kubernetes==28.1.0
langchain==0.0.354
langchain-community==0.0.9
langchain-core==0.1.7
langsmith==0.0.77
lxml==5.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pathtools==0.1.2
peft @ git+https://github.com/huggingface/peft.git@7120c9a2636f93c8db3fc4ae466e02f338ecead8
posthog==3.1.0
protobuf==4.25.1
psutil==5.9.7
pulsar-client==3.4.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.17.2
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rsa==4.9
safetensors==0.4.1
schema==0.7.5
scikit-learn==1.2.2
scipy==1.11.4
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@4ab5fb8941a38d172b3883c152c34ae2a0b83a68
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
vine==5.1.0
wandb==0.15.3
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
